create two separate components for avatar utils
