+++ .bitmap
@@ -1,6 +1,7 @@
 /* THIS IS A BIT-AUTO-GENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. */
 
 {
+  "string/get-initials": {"mainFile":"index.ts","rootDir":"scopes/toolbox/string/get-initials","exported":false},
   "teambit.code/ui/code-tab-page@0.0.283": {"mainFile":"index.ts","rootDir":"scopes/code/ui/code-tab-page"},
   "teambit.code/ui/code-tab-tree@0.0.283": {"mainFile":"index.ts","rootDir":"scopes/code/ui/code-tab-tree"},
   "teambit.code/ui/code-view@0.0.282": {"mainFile":"index.ts","rootDir":"scopes/code/ui/code-view"},
@@ -188,5 +189,6 @@
   "teambit.workspace/ui/workspace-component-card@0.0.15": {"mainFile":"index.ts","rootDir":"scopes/workspace/ui/workspace-component-card"},
   "teambit.workspace/variants@0.0.283": {"mainFile":"index.ts","rootDir":"scopes/workspace/variants"},
   "teambit.workspace/workspace@0.0.294": {"mainFile":"index.ts","rootDir":"scopes/workspace/workspace"},
+  "url/add-avatar-query-params": {"mainFile":"index.ts","rootDir":"scopes/toolbox/url/add-avatar-query-params","exported":false},
   "version": "14.8.9-dev.298"
 }
