+++ scopes/design/ui/Avatar/org-avatar.tsx
@@ -1,11 +1,8 @@
-import classNames from 'classnames';
 import React, { PureComponent } from 'react';
-
+import classNames from 'classnames';
+import { addAvatarQueryParams } from '@teambit/url.add-avatar-query-params';
 import { AccountObj } from './avatar';
-// import Tooltip from 'react-tooltip';
 import styles from './styles.module.scss';
-import { addQueryParams } from './utils';
-// import { v1 } from 'uuid';
 
 type Props = {
   account: AccountObj;
@@ -14,42 +11,19 @@ type Props = {
   fontSize?: number;
   className?: string;
   imgClassName?: string;
-  // hideTooltip?: boolean;
 };
 
 export class OrgAvatar extends PureComponent<Props> {
-  // state = {
-  //   tooltipId: null
-  // };
-
-  // componentDidMount() {
-  // 	//mount only happens in client side
-  // 	//setting state here will prevent id reallocated after serverside rendering
-  // 	//prevent double render when id change in clinet side after serverside rendering
-  // 	this.setState({ tooltipId: v1() });
-  // }
-
   render() {
-    const {
-      account,
-      size,
-      imageSize = size,
-      fontSize = size * 0.35,
-      className,
-      imgClassName,
-      // hideTooltip = false
-    } = this.props;
-    // const { tooltipId } = this.state;
+    const { account, size, imageSize = size, fontSize = size * 0.35, className, imgClassName } = this.props;
 
-    const { profileImage = '' /* , name = '', displayName = '' */ } = account;
-    const profileImageWithParams = addQueryParams(profileImage, imageSize);
+    const { profileImage = '' } = account;
+    const profileImageWithParams = addAvatarQueryParams(profileImage, imageSize, styles.defaultAvatarBgColor);
 
     return (
       <div
         className={classNames(styles.default, styles.avatar, className)}
         style={{ width: `${size}px`, height: `${size}px` }}
-        // data-for={tooltipId}
-        // data-tip={displayName || name}
       >
         {profileImageWithParams && (
           <img src={profileImageWithParams} className={classNames(styles.avatarImg, imgClassName)} />
@@ -57,16 +31,6 @@ export class OrgAvatar extends PureComponent<Props> {
         <span className={styles.defaultAvatar}>
           <i className="bitcon-org" style={{ fontSize: `${fontSize}px`, lineHeight: `${size}px` }} />
         </span>
-        {/* {tooltipId && (
-					<Tooltip
-						className={styles.tooltip}
-						id={tooltipId}
-						place="bottom"
-						type="dark"
-						effect="solid"
-						disable={hideTooltip}
-					/>
-				)} */}
       </div>
     );
   }
