+++ scopes/toolbox/string/capitalize/capitalize.docs.md
@@ -1,0 +1,13 @@
+---
+labels: ['typescript', 'utils', 'string', 'text', 'capitalize']
+description: 'Capitalize a string.'
+---
+
+import { capitalize } from './capitalize';
+
+Capitalize the first letter of a string.  
+For example:
+
+```js live
+capitalize('scope name');
+```
