+++ scopes/toolbox/string/capitalize/capitalize.spec.ts
@@ -2,8 +2,6 @@ import { capitalize } from './capitalize';
 
 describe('capitalize()', () => {
   it('should capitalize a single word', () => {
-    // @ts-ignore
-    // eslint-disable-next-line no-undef
     expect(capitalize('foo')).toEqual('Foo');
   });
 });
