+++ scopes/toolbox/string/get-initials/get-initials.docs.md
@@ -1,0 +1,13 @@
+---
+labels: ['typescript', 'utils', 'string', 'text']
+description: 'Get initials of a string.'
+---
+
+import { getInitials } from './get-initials';
+
+A util function that return the initials of a string if possible.
+For example:
+
+```js live
+getInitials('scope name');
+```
