+++ scopes/toolbox/string/get-initials/index.ts
@@ -1,0 +1,1 @@
+export { getInitials } from './get-initials';
