+++ scopes/toolbox/url/add-avatar-query-params/add-avatar-query-params.docs.md
@@ -1,0 +1,9 @@
+---
+labels: ['typescript', 'utils', 'url', 'images', 'avatar']
+description: 'Build image url with params.'
+---
+
+Adds query params to avatar image, according to the type of the hosting service.
+
+- For Gravatar, adds size
+- For ImgIX, adds size, default background, and scaling method
