+++ scopes/toolbox/url/add-avatar-query-params/index.ts
@@ -1,0 +1,1 @@
+export { addAvatarQueryParams } from './add-avatar-query-params';
